here the float entry is accepted in days

+++ backend/app/config.py
@@ -28,7 +28,7 @@ class Settings(BaseSettings):
     config_pin: str = "1234"
 
     # Data Retention Settings (Time-based FIFO cleanup)
-    data_retention_days: int = 90         # Keep last 90 days (3 months) of data
+    data_retention_days: float = 90       # Keep last 90 days (3 months) of data - supports decimals (e.g., 0.0833 = 2 hours)
     data_retention_max_rows: int = 5000000  # Safety limit - max rows in database
     cleanup_hour: int = 2                 # Daily cleanup runs at this hour (24-hour format)
